Add parsing tests and fix bug
When adding a property test for parsing commands
(Props.ShellRun.Parsing.Commands), it was discovered that there was a flaw
in the parsing logic. If a user provided a legend file and requested a command
that has a comma in it (e.g. 'echo ,'), this will be treated as multiple
commands because we are naively splitting on the ','.

This bug can be witnessed by running:

  shell-run --legend=legend.txt "echo ,"

and noticing that two commands are run: 'echo ' and ''. The main logic error
is:

  1. If a legend file is present, split all cli commands _before_
     checking if it is a key first.

One is fixed by only splitting the results of successful key lookups, not
every command encountered (e.g. CLI args).

The other flaw is using a single comma as a delimiter in the legend file.
This decision means one cannot include commas in any commands defined in
the legend file (e.g., 'key=cmd1,cmd2,with,commas,cmd3'), which is unfortunate
because it is a valid bash character. Thus, we instead use a double comma
('key=cmd1,,cmd2,with,commas,,cmd3'), hoping this does not cause many issues.