Improve functional test output
By swapping the beforeAll/afterAll functions for before/After, the test output
actually seems to show each test as it should. The downside is we are now
running startup/cleanup before/after each spec, though this doesn't seem to
impact perfomance at all. This whole setup is super fragile, because we are
running tests that capture stdout, while the test suite is also printing to
stdout.

At some point maybe this could be refactored, though the application
necessarily writes to stdout, so getting around these difficulties may not
be easy.