Async updates
Previously, I was combining all commands into an asynchronous task via
`mapConcurrently` and `async`. I was then manually keeping track of the async
status while running the counter, cancelling the commands if we timed out.

For reasons that are still not entirely clear to me, if an exception occurred
in one of the command threads (not in the command itself, but rather a
Haskell IO exception _around_ the command, e.g. reading a stale File
Descriptor with lazy IO), this error would be swallowed. The error would
manifest by the entire program exiting (`mapConcurrently` kills everything
when an exception occurs), which is fine, but the exception wouldn't be
re-thrown. `mapConcurrently` _does_ rethrow exceptions, so this error was
definitely somewhere in my logic, but I did not discover where.

I have since switched to using Async's `race_` function to manage the timer
and commands tasks. The means I no longer have to manually kill the async
tasks if we timeout (`race` automatically cancels the 'loser' of the two
actions). Furthermore, errors are now detectable in the main `runCommands`
thread via `Except.try` (previously this would fail to discover exceptions).
I don't know _why_ this is better, other than concurrency is hard, and
it's better to let Async handle as much of the logic as possible, rather than
re-implementing it myself ;-).