Fix cabal ci
The functional tests were failing because they are not thread-safe, and
it was being built with '-with-rtsopts="-N"'. Removing this from the
global ghc-options fixes the test.

I also moved -threaded to the exe specific ghc-options since that's the
only place it matters.

Finally, optimization was moved to the top-level in the cabal project
file because it did not appear to be picked up when it was under the
package.