Fundamental changes to logging
Command logging has long been a pain point. Threads interfere with
each other, logs get overwritten, and -- the bane of my existence
-- buffering. The changes here are meant to address these
shortcomings.

First, rather than each thread handling its own printing directly,
we send logs to a 'log queue', then we have a dedicated thread
that reads the queue and prints logs as they come. This does not
directly solve the above issues, but it simplifies the logic.

Second, for sub-logging, we use a pseudo-terminal (pty) to communicate
with the child process and receive its logs to add to the queue. The
reason for the pty is to overcome buffer difficulties. The previous
strategy of passing the `CreatePipe` option simply would not stream the
child's stdout if the child was not flushing output. Attempts to
override the buffering mode were unsuccessful. Happily, the pty solves
this nicely.

Third, for sub-logging, we combine stdout and stderr into a single
handle and treat it the same (i.e. as general 'output'). Many bash
scripts treat them the same (e.g. `echo ... >&2`) thus we lose the
ability to differentiate them anyway.

Finally, prior to these changes, there was complicated logic for
determining if we need to 'clear' the current line. Different logs would
clear and/or update the carriage returns. With multiple threads running
this became impossible to do correctly. Instead, _every_ line 'clears'
the current line in case there is any 'leftover' text. Additionally,
logs are now responsible for knowing their own newline/carriage return
behavior. For instance, lines that should be overwritten (e.g. the
running timer) include a carriage return. This way a partciular log
does not need to try to understand the current 'logging context'
(i.e. what the last log was).