Fix functional tests
It was nice having the tests verify the pretty colors, but
it was always handing by a thread as it required exact text
matches -- necessitating consistent runtimes and (gulp)
shell portability. I _thought_ this might be okay and it took
exactly one ci run on a different machine to disabuse me of
that notion :-). Specifically, the part of the text verifications
that failed was an error looking for '/bin/sh'.

That has been removed to do simpler text checks that should be far
more robust.