Give up on cabal+doctest+ci for now
Basically, doctest+cabal is hard. For reasons I don't understand,
doctest started failing on ci for cabal builds only (both Nix and
Stack work fine because #reasons). The error is complaining that it
can't find the System.Console.Pretty module (from pretty-terminal)
in ShellRun.Logging.Log.hs, but I have no idea why it's looking in
there in the first place.

Doctest doesn't compile the project, and that file shouldn't be
included in any of the files I'm running doctest on anyway.
Whatever the reason, my attempts to fix this have been in vain,
so for now it's being disabled on CI.