Add file logging option
There is now an option to log to a file. Regular logging is unaffected
by this flag; it is merely an additional option.

Rather than manually making sure every log statement is accompanied by
a logFile statement, it was clear that there would be an advantage to
making the "if option then log to file" logic part of the RegionLogger
itself. This way we are guaranteed to have the correct behavior (i.e.,
cannot miss any statements).

However, when implementing this, it meant that we have to thread our
filepath argument through a bunch of different functions. Thus it became
preferable to instead move the RegionLogger instances onto ShellT instead
of IO. This way we'd have access to our ReaderT env. Additionally, this
necessitated moving the tryTimeSh and tryTimeShRegion into ShellT rather
than plain IO (since the latter needs the file log logic). Thus all of Async
and some of the IO logic was moved directly into the new ShellT module.

Furthermore, the log queue was brought back. This time it is not used for
general logging, only file logging, since that must be atomic. To sum up, when
ShellT's RegionLogger functions are called, we check if the log file options
is set. If so, the log gets sent to the queue, and a separate thread
periodically writes these logs to a file.

Finally, we mitigate the stubborn "IOException reading a file" bug. The new file log
shows that it is not totally gone, but by changing the close_fds value to False, we
definitely fix some of the prior bad reads."