Change logging semantics
In a desire to keep the API relatively simple and consistent, we have changed
how the command/file logging works. File logging is no longer tied to console
logging. We have:

- Logging commands in the console is controlled by --cmd-log.
- File logging is controlled by --file-log. File logging _always_ streams
  command logs.
- Options that affect how logs are displayed (--key-show, --cmd-truncate) affects
  the console logs only. File logging always shows the full command.

To accomplish this, we have had to add a new IO function for running the commands.
We currently have:
  - tryTimeSh (no command logging at all)
  - tryTimeShRegion (stream command logging to the console)
  - tryTimeShNoRegion (new, stream command logging to a file)

We have to modify our logging logic so that the Log type no longer contains the
formatted text (e.g. prefix, color info, command name). Formatting needs to happen
in the actual printing/queue writing, since we want access to ReaderT vars that
control the formatting.

We also add a concept of destination to the Log type. Most of the time this will
be both (Console and File). But there are exceptions:
  - --file-log but no --cmd-log means streamed logs should be sent to the file queue
    only.
  - Running timer logs should not be sent to the file.