Swap screenshots for html formatting
The screenshots were helpful but didn't render well on mobile,
and they're a massive pain both to take and update.

With actual text, we can easily update them in the future, and
we no longer have any rendering problems. The html formatting
is a bit tricky, but it's still easier than taking/editing
screenshots.