Prevent functional tests stdout interference
It turns out, the reason the functional tests interfered with stdout
was due to concurrent-output's usage. The solution is to abstract all such
usages to a typeclass, so our functional tests can have mocked logic.
We can do this for the low cost of an extra type parameter on the Logging
type.