Fix console log 'bug'
We actually were not dealing with a concurrency bug regarding our
output (mercifully); in reality, the issue is that the 'final' message
we receive when we encounter an error is inconsistent. This was causing
our tests to fail and made us mistakenly believe we were dropping
messages.

Of course this isn't totally satisfactory, but it is much less of an issue
than one our core logging infra being unreliable.