Fix queue read/write atomicity
The previous "queue reads + writes are atomic" attempt made no sense.
The problem was an async exception killing our thread after a read but
before a write. That has nothing to do with semaphores blocking other
threads from performing multiple reads. Furthermore, we only had one
thread running that code anyway.

Somehow, it appeared to 'work' (i.e. we were no longer losing messages).
Maybe the added lock/unlock behavior caused enough of a performance delay
that the thread was more likely to be killed during that, rather than a
read+write? Who knows.

In any case, the correct solution is to temporarily block async exceptions
_inside_ the loop, so a read will be accompanied by the corresponding write.
Because the mask is inside the loop, the thread can still be correctly
cancelled (in between loops, essentially). The only danger is if an action
inside the loop blocks for a long time, though the only thing we are doing
is reading from a queue and writing to stdout/file, which is probably
safe.