Consolidate types, parsing
We combine our disparate Args/Toml types into single types.
We then use the "higher-kinded data" approach to "evolve" the data
across phases. This allows us to reuse the same types (and fields)
and combine Args + Toml in a reasonably principled way.

This has the happy side-affect of greatly reducing the Args type /
parsing. Instead of a gigantic type and many parsers, we have
nested fields for better organization and a generic "WithDisable"
field to represent the disable flags.

We also switch some tests to use hedgehog purely for its
pretty-printing capabilities: it diffs records very nicely, which
is crucial in investigating assertion failures for complicated
data.