Move 'log-file-size-mode nothing' test to integration
It turns out, 'log-file-size-mode' tests already exist in the
integration suite. This makes sense as the functional test recently
added does not actually run shrun. Thus it makes sense to move this
test to the integration test suite with the others.

In the future, we can add real ones to the functional suite, provided
we actually run shrun (then we can test that the new log file really
is > than the previous one, for example).