Rename cmd-log-size to cmd-log-read-size
We rename this parameter to make it more explicit that this param
corresponds to the amount of bytes _read_ from the command. This could
potentially be confused with the max number of bytes _displayed_ in
a command log.

These are mostly the same at the moment (module a small constant
factor for the '[Command]...' prefix), but it is possible we could
offer both in the future (the hypothetical latter param is similar but
distinct from --cmd-log-line-trunc). Hence we make the read size
clearer.

We also reuse some default params in the configuration.