Extensive option renaming/reorganizing
The logging related options were a bit o a mess, as the names did
did little to describe where the options were relevant ("does
--cmd-name-trunc affect non-command console logs? What about
key-hide?").

Furthermore, one of the names was actually "wrong":
--cmd-log-strip-control implies it's the strip-control concept applied
to (console) "command logs", but it's actually applied to _all_
console logs.

Thus we come up with a consistent naming scheme which greatly clarifies
things, at the cost of some verbosity. That is, we now have prefixes
(and toml):

    - `--log-*` (`log`): common logging configuration.
    - `--cmd-log-*` (`cmd-log`): applies to _command logs_ everywhere.
      (i.e. console and file).
    - `--console-log-*` (`console-log`): applies to console logs.
    - `--file-log-*` (`file-log`): applies to file logs.