Improve log formatting
The main visible change here is making line truncation more
consistent. We apply line truncation __all the time__ (to console logs)
when it is active i.e. no longer dependent on presence of a command.

We also simplify the console/file log formatting s.t. most of the
shared logic is in a single function, making it easier to maintain
consistency between the two. It is much easier to see the difference
between the two now (timestamp vs. colorization, mostly).

Finally, we simplify the formatting tests. Previously, the formatting
tests were quite weak as we tried to stuff too much generality into
property tests. That is, we generated arbitrary logs then tried to
assert properties. But this was quite hard as the properties are very
weak!

Instead, we randomize just a few values on the logs and make simple,
robust assertions.