Revert "Make flaky OSX buffering test more robust"
This reverts commit 2f0ea39a69a13f65cdc1c036090d1378d46e18d6.

Turns out, 2f0ea39a69a13f65cdc1c036090d1378d46e18d6 was not enough
to make the tests robust. The problem is that the order of the last
few logs is not terribly reliable, especially on OSX.

Instead, we give up on asserting any sort of order for them and instead
merely look for existence. This should be more reliable, but note that
we still require the first few logs to be ordered deterministically
(morally).