Add Default class for default values
Instead of having functions e.g. defaultPollInterval we instead
create a typeclass @class Default a where def :: a@.

We do this to stress the canonicity of such defaults. There really
should only be one default for each type. We previously had a bug
(StripControl) where we had inconsistent defaults, due to a default
value being manually set in multiple locations. By using a typeclass,
we ensure this does not happen.

We also utilize the Default class to significantly simplify the
Args + Toml merge logic. Technically this is orthogonal to the class,
though the class makes it easier to write some convenient
operators / functions.