Disable 'block-line-buffer' strategy less often
3a77b3c05595c42570cec4ff154e7c66c358829e was bit overzealous in
disabling the 'block-line-buffer' strategy. In fact, this strategy
is perfectly reasonable with multiple commands, as long as file-logging
is not active.

Thus we only throw errors for it is active, along with multiple
commands and file-logging.

That said, it is an open question whether it is the right __default__
for multiple commands (console-logging only). The 'block-line-buffer'
strategy gets most of its mileage with file-logging, as broken logs
matter less for console logs.

Nevertheless we default to 'block-line-buffer' in these situations, as
it doesn't appear harmful, and this preserves the nice symmetry of

    'block-line-buffer' enabled by default
      <=> 'block-line-buffer' is allowed

Note that this means manually setting '--command-log-read-strategy'
to 'block-line-buffer' is never necessary, as shrun will pick this
by default in exactly the cases where it is allowed.