Remove local INLINEABLEs
We annotate monad-polymorphic functions with INLINEABLE, so that
these functions can potentially be specialized, allowing optimizations.

Previously, we also annotated local functions (i.e. those within
a top-level INLINEABLE function) with INLINEABLE, out of paranoia.
But there appears to be no evidence this is beneficial. Benchmarks
here do not show a difference, and examining the core in simple
cases does not show a difference either.

Therefore, for readability/maintenance reasons, we remove these
seemingly useless INLINEABLEs.