Implement command dependencies
Introduces the --command-graph option, which allows commands to be
run sequentially. Each command is given an index based on its
left-to-right order. For instance:

  --command-graph '1 -> 3, 2 -> 3'

will block the 3rd command from starting until 1 and 2 have finished
successfully. If either fail, 3 will not be run.

Note that indexes are assigned _after_ legend expansion.