Add ProcessHandle to running command status.
Adds ProcessHandle to CommandRunning status for later cancellation, to
fix the linux CI bug and hopefully make cancellation more robust.
To facilitate this, we remove the "no stream" IO functions as we always
need to manually create a process, to acquire the handle.

This is arguably a good idea anyway, as it means more logic is shared
regardless of command/file logs (previously we had an issue where
subprocessses were only cancelled with command logging).