Add cancellation fallback
When shrun is terminated, we want all currently running subcommands
to be terminated. We rely on exceptions' automatic proprogation via
async and process, which generally works pretty well. There is at least
one scenario where it does not work: CI linux

Consider the command 'shrun "sleep 10"'. This will spawn 3 commands:

- shrun "sleep 10"
- /bin/sh -c "sleep 10"
- "sleep 10"

In other words, each command given to shrun actually spawns 2 commands,
a shell wrapper which spawns the command itself. Note that behavior is
platform dependent e.g. on my linux machine the '/bin/sh' command
terminates immediately, leaving only the 'sleep 10' command.

Now suppose shrun is cancelled e.g. it receives SIGINT or SIGTERM.
The async thread that is blocked on /bin/sh will receive the exception,
which is then sent to the process. On CI osx, this exception is
proprogated to the underlying 'sleep 10' command, so everything is
cancelled. On CI linux, however, this does _not_ happen, so while the
/bin/sh command is cancelled, 'sleep 10' is left running. This is
not what we want.

Hence we add extra logic that tries to manually end the process.